TASK 3.1.2 (исправил замечания)
