+++ src/main/java/ru/itmentor/spring/boot_security/demo/configs/MvcConfig.java
@@ -8,7 +8,5 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class MvcConfig implements WebMvcConfigurer {
     public void addViewControllers(ViewControllerRegistry registry) {
-//        registry.addViewController("/user").setViewName("user");
-//        registry.addViewController("/login").setViewName("login");
     }
 }
