+++ src/main/java/ru/itmentor/spring/boot_security/demo/controllers/AuthController.java
@@ -8,25 +8,24 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import ru.itmentor.spring.boot_security.demo.models.User;
-import ru.itmentor.spring.boot_security.demo.services.RegistrationService;
 import ru.itmentor.spring.boot_security.demo.services.RoleServices;
+import ru.itmentor.spring.boot_security.demo.services.UserServices;
 import ru.itmentor.spring.boot_security.demo.util.UserValidator;
 
 
 @Controller
 public class AuthController {
 
-    private final RegistrationService registrationService;
     private final UserValidator userValidator;
     private final RoleServices roleServices;
+    private final UserServices userServices;
 
     @Autowired
-    public AuthController(RegistrationService registrationService, UserValidator userValidator, RoleServices roleServices) {
-        this.registrationService = registrationService;
+    public AuthController(UserValidator userValidator, RoleServices roleServices, UserServices userServices) {
         this.userValidator = userValidator;
         this.roleServices = roleServices;
+        this.userServices = userServices;
     }
 
     @GetMapping("/login")
@@ -46,7 +45,7 @@ public class AuthController {
         if (bindingResult.hasErrors()) {
             return "registration";
         }
-        registrationService.register(user);
+        userServices.register(user);
         return "redirect:/login";
     }
 }
