--- src/main/java/ru/itmentor/spring/boot_security/demo/services/RegistrationService.java
@@ -1,7 +1,0 @@
-package ru.itmentor.spring.boot_security.demo.services;
-
-import ru.itmentor.spring.boot_security.demo.models.User;
-
-public interface RegistrationService {
-    void register(User User);
-}
